requirements.txt
asgiref==3.8.1
black==24.8.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==2.0.12
click==8.1.7
colorama==0.4.6
cssbeautifier==1.15.1
cssselect2==0.7.0
Django==5.1
django-bootstrap5==24.2
django-ckeditor-5==0.2.13
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-js-asset==2.2.0
django-json-widget==2.0.1
django-mptt==0.16.0
django-polymorphic==3.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djlint==1.34.1
EditorConfig==0.12.4
future==1.0.0
html-tag-names==0.1.2
html-void-elements==0.1.0
iamport==0.0.1
iamport-rest-client==0.9.0
idna==3.8
jsbeautifier==1.15.1
json5==0.9.25
lxml==5.3.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
psycopg2-binary==2.9.9
PyJWT==2.9.0
PyYAML==6.0.2
regex==2023.12.25
reportlab==4.2.2
requests==2.31.0
serializers==0.2.4
six==1.16.0
sorl-thumbnail==12.10.0
sqlparse==0.5.1
svglib==1.5.1
tinycss2==1.3.0
tomli==2.0.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.20
webencodings==0.5.1
